+++ modules/module-mongodb-storage/src/migrations/db/migrations/1752661449910-connection-reporting.ts
@@ -1,0 +1,24 @@
+import { migrations } from '@powersync/service-core';
+import * as storage from '../../../storage/storage-index.js';
+import { MongoStorageConfig } from '../../../types/types.js';
+
+export const up: migrations.PowerSyncMigrationFunction = async (context) => {
+  const {
+    service_context: { configuration }
+  } = context;
+  // No-op - moved to 1762790715147-connection-reporting2
+};
+
+export const down: migrations.PowerSyncMigrationFunction = async (context) => {
+  const {
+    service_context: { configuration }
+  } = context;
+
+  const db = storage.createPowerSyncMongo(configuration.storage as MongoStorageConfig);
+
+  try {
+    await db.db.dropCollection('connection_report_events');
+  } finally {
+    await db.client.close();
+  }
+};