+++ modules/module-postgres-storage/src/utils/test-utils.ts
@@ -1,7 +1,8 @@
 import { framework, PowerSyncMigrationManager, ServiceContext, TestStorageOptions } from '@powersync/service-core';
 import { PostgresMigrationAgent } from '../migrations/PostgresMigrationAgent.js';
 import { normalizePostgresStorageConfig, PostgresStorageConfigDecoded } from '../types/types.js';
-import { PostgresBucketStorageFactory } from './PostgresBucketStorageFactory.js';
+import { PostgresReportStorage } from '../storage/PostgresReportStorage.js';
+import { PostgresBucketStorageFactory } from '../storage/PostgresBucketStorageFactory.js';
 
 export type PostgresTestStorageOptions = {
   url: string;
@@ -12,7 +13,7 @@ export type PostgresTestStorageOptions = {
   migrationAgent?: (config: PostgresStorageConfigDecoded) => PostgresMigrationAgent;
 };
 
-export const postgresTestSetup = (factoryOptions: PostgresTestStorageOptions) => {
+export function postgresTestSetup(factoryOptions: PostgresTestStorageOptions) {
   const BASE_CONFIG = {
     type: 'postgresql' as const,
     uri: factoryOptions.url,
@@ -48,6 +49,21 @@ export const postgresTestSetup = (factoryOptions: PostgresTestStorageOptions) =>
   };
 
   return {
+    reportFactory: async (options?: TestStorageOptions) => {
+      try {
+        if (!options?.doNotClear) {
+          await migrate(framework.migrations.Direction.Up);
+        }
+
+        return new PostgresReportStorage({
+          config: TEST_CONNECTION_OPTIONS
+        });
+      } catch (ex) {
+        // Vitest does not display these errors nicely when using the `await using` syntx
+        console.error(ex, ex.cause);
+        throw ex;
+      }
+    },
     factory: async (options?: TestStorageOptions) => {
       try {
         if (!options?.doNotClear) {
@@ -66,8 +82,8 @@ export const postgresTestSetup = (factoryOptions: PostgresTestStorageOptions) =>
     },
     migrate
   };
-};
+}
 
-export const PostgresTestStorageFactoryGenerator = (factoryOptions: PostgresTestStorageOptions) => {
+export function postgresTestStorageFactoryGenerator(factoryOptions: PostgresTestStorageOptions) {
   return postgresTestSetup(factoryOptions).factory;
-};
+}