+++ modules/module-postgres-storage/src/utils/utils-index.ts
@@ -2,3 +2,4 @@ export * from './bson.js';
 export * from './bucket-data.js';
 export * from './db.js';
 export * from './ts-codec.js';
+export * as test_utils from './test-utils.js';