+++ modules/module-postgres/test/src/storage_combination.test.ts
@@ -7,7 +7,7 @@ describe.skipIf(!env.TEST_POSTGRES_STORAGE)('replication storage combination - p
   test('should allow the same Postgres cluster to be used for data and storage', async () => {
     // Use the same cluster for the storage as the data source
     await using context = await WalStreamTestContext.open(
-      postgres_storage.PostgresTestStorageFactoryGenerator({
+      postgres_storage.test_utils.postgresTestStorageFactoryGenerator({
         url: env.PG_TEST_URL
       }),
       { doNotClear: false }