+++ packages/service-core/src/routes/endpoints/sync-stream.ts
@@ -7,8 +7,8 @@ import * as util from '../../util/util-index.js';
 
 import { authUser } from '../auth.js';
 import { routeDefinition } from '../router.js';
+import { APIMetric, event_types } from '@powersync/service-types';
 
-import { APIMetric } from '@powersync/service-types';
 import { maybeCompressResponseStream } from '../compression.js';
 
 export enum SyncRoutes {
@@ -25,7 +25,7 @@ export const syncStreamed = routeDefinition({
   authorize: authUser,
   validator: schema.createTsCodecValidator(util.StreamingSyncRequest, { allowAdditional: true }),
   handler: async (payload) => {
-    const { service_context, logger } = payload.context;
+    const { service_context, logger, token_payload } = payload.context;
     const { routerEngine, storageEngine, metricsEngine, syncContext } = service_context;
     const headers = payload.request.headers;
     const userAgent = headers['x-user-agent'] ?? headers['user-agent'];
@@ -43,6 +43,14 @@ export const syncStreamed = routeDefinition({
       client_id: clientId,
       user_id: payload.context.user_id,
       bson: useBson
+    };
+    const sdkData: event_types.ConnectedUserData & event_types.ClientConnectionEventData = {
+      client_id: clientId ?? '',
+      user_id: payload.context.user_id!,
+      user_agent: userAgent as string,
+      // At this point the token_payload is guaranteed to be present
+      jwt_exp: new Date(token_payload!.exp * 1000),
+      connected_at: new Date(streamStart)
     };
 
     if (routerEngine.closed) {
@@ -69,6 +77,7 @@ export const syncStreamed = routeDefinition({
     const tracker = new sync.RequestTracker(metricsEngine);
     try {
       metricsEngine.getUpDownCounter(APIMetric.CONCURRENT_CONNECTIONS).add(1);
+      service_context.eventsEngine.emit(event_types.EventsEngineEventType.SDK_CONNECT_EVENT, sdkData);
       const syncLines = sync.streamResponse({
         syncContext: syncContext,
         bucketStorage,
@@ -134,6 +143,10 @@ export const syncStreamed = routeDefinition({
           }
           controller.abort();
           metricsEngine.getUpDownCounter(APIMetric.CONCURRENT_CONNECTIONS).add(-1);
+          service_context.eventsEngine.emit(event_types.EventsEngineEventType.SDK_DISCONNECT_EVENT, {
+            ...sdkData,
+            disconnected_at: new Date()
+          });
           logger.info(`Sync stream complete`, {
             ...tracker.getLogMeta(),
             stream_ms: Date.now() - streamStart,
@@ -144,6 +157,10 @@ export const syncStreamed = routeDefinition({
     } catch (ex) {
       controller.abort();
       metricsEngine.getUpDownCounter(APIMetric.CONCURRENT_CONNECTIONS).add(-1);
+      service_context.eventsEngine.emit(event_types.EventsEngineEventType.SDK_DISCONNECT_EVENT, {
+        ...sdkData,
+        disconnected_at: new Date()
+      });
     }
   }
 });