+++ modules/module-core/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @powersync/service-module-core
 
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [b364581]
+- Updated dependencies [d889219]
+- Updated dependencies [0ace0d3]
+- Updated dependencies [7eb7957]
+- Updated dependencies [b364581]
+  - @powersync/service-core@1.16.2
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [c6bdb4f]
+  - @powersync/service-core@1.16.1
+
+## 0.2.9
+
+### Patch Changes
+
+- 80fd68b: Add SDK usage reporting support.
+- 88982d9: Migrate to trusted publishing
+- Updated dependencies [80fd68b]
+- Updated dependencies [88982d9]
+  - @powersync/service-core@1.16.0
+  - @powersync/lib-services-framework@0.7.9
+  - @powersync/service-types@0.13.1
+  - @powersync/service-rsocket-router@0.2.6
+
 ## 0.2.8
 
 ### Patch Changes