+++ service/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @powersync/service-image
 
+## 1.16.2
+
+### Patch Changes
+
+- d889219: Fix memory leaks when retrying replication after errors.
+- 7eb7957: [Postgres] Remove usage of pg_logical_slot_peek_binary_changes due to performance issues in some cases
+- Updated dependencies [b364581]
+- Updated dependencies [d889219]
+- Updated dependencies [0ace0d3]
+- Updated dependencies [7eb7957]
+- Updated dependencies [b364581]
+  - @powersync/service-module-mongodb-storage@0.12.11
+  - @powersync/service-core@1.16.2
+  - @powersync/service-module-postgres-storage@0.10.11
+  - @powersync/service-module-postgres@0.16.11
+  - @powersync/service-module-mongodb@0.12.11
+  - @powersync/service-module-mysql@0.9.11
+  - @powersync/service-module-core@0.2.11
+
+## 1.16.1
+
+### Patch Changes
+
+- c6bdb4f: [MongoDB storage] Fix migration for indexes on connection_report_events.
+- Updated dependencies [c6bdb4f]
+  - @powersync/service-module-mongodb-storage@0.12.10
+  - @powersync/service-core@1.16.1
+  - @powersync/service-module-mongodb@0.12.10
+  - @powersync/service-module-mysql@0.9.10
+  - @powersync/service-module-postgres@0.16.10
+  - @powersync/service-module-core@0.2.10
+  - @powersync/service-module-postgres-storage@0.10.10
+
+## 1.16.0
+
+### Minor Changes
+
+- 80fd68b: Add SDK usage reporting support.
+
+### Patch Changes
+
+- e465ba7: Node 22.21.1, with support for --max-old-space-size-percentage.
+- Updated dependencies [0e9aa94]
+- Updated dependencies [80fd68b]
+- Updated dependencies [88982d9]
+  - @powersync/service-module-postgres@0.16.9
+  - @powersync/service-module-postgres-storage@0.10.9
+  - @powersync/service-module-mongodb-storage@0.12.9
+  - @powersync/service-module-mongodb@0.12.9
+  - @powersync/service-core@1.16.0
+  - @powersync/service-module-mysql@0.9.9
+  - @powersync/service-module-core@0.2.9
+  - @powersync/lib-services-framework@0.7.9
+  - @powersync/service-rsocket-router@0.2.6
+
 ## 1.15.8
 
 ### Patch Changes